move templates to bokeh.core
