+++ bokeh/core/_templates/autoload_js.js
