+++ bokeh/core/_templates/css_resources.html
