+++ bokeh/core/_templates/doc_js.js
