+++ bokeh/core/_templates/js_resources.html
