+++ bokeh/core/_templates/notebook_load.html
