+++ bokeh/core/_templates/script_tag.html
