+++ bokeh/core/templates.py
@@ -1,23 +1,23 @@
 ''' The templates module contains Jinja2 templates used by Bokeh embed
 Bokeh models (e.g. plots, widgets, layouts) in various ways.
 
-.. bokeh-jinja:: bokeh.templates.AUTOLOAD_JS
-.. bokeh-jinja:: bokeh.templates.AUTOLOAD_TAG
-.. bokeh-jinja:: bokeh.templates.CSS_RESOURCES
-.. bokeh-jinja:: bokeh.templates.DOC_JS
-.. bokeh-jinja:: bokeh.templates.FILE
-.. bokeh-jinja:: bokeh.templates.JS_RESOURCES
-.. bokeh-jinja:: bokeh.templates.NOTEBOOK_LOAD
-.. bokeh-jinja:: bokeh.templates.NOTEBOOK_DIV
-.. bokeh-jinja:: bokeh.templates.PLOT_DIV
-.. bokeh-jinja:: bokeh.templates.SCRIPT_TAG
+.. bokeh-jinja:: bokeh.core.templates.AUTOLOAD_JS
+.. bokeh-jinja:: bokeh.core.templates.AUTOLOAD_TAG
+.. bokeh-jinja:: bokeh.core.templates.CSS_RESOURCES
+.. bokeh-jinja:: bokeh.core.templates.DOC_JS
+.. bokeh-jinja:: bokeh.core.templates.FILE
+.. bokeh-jinja:: bokeh.core.templates.JS_RESOURCES
+.. bokeh-jinja:: bokeh.core.templates.NOTEBOOK_LOAD
+.. bokeh-jinja:: bokeh.core.templates.NOTEBOOK_DIV
+.. bokeh-jinja:: bokeh.core.templates.PLOT_DIV
+.. bokeh-jinja:: bokeh.core.templates.SCRIPT_TAG
 
 '''
 from __future__ import absolute_import
 
 from jinja2 import Environment, PackageLoader
 
-_env = Environment(loader=PackageLoader('bokeh', '_templates'))
+_env = Environment(loader=PackageLoader('bokeh.core', '_templates'))
 
 JS_RESOURCES = _env.get_template("js_resources.html")
 CSS_RESOURCES = _env.get_template("css_resources.html")
